data: j721e: Add json object to describe dependencies
Describe the utype string to be used for each resource
"restrictHosts" describes an array of hosts where this
resource should be allocated. Give a warning otherwise

"autoAllocStart" describes if the start range should be
automatically allocated for this resource. When set to
false, solver will not try to auto assign this.
Warnings will be generated in case there are overlapping
entries for this resource.

"copyFromUtype" describes another utype to copy the data
resource allocation from

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>