data: j721e: Describe depdendencies for block copy channels
Introuce a new property "blockCopy" for each resource utype
When set to true, the resource group  will have an additional
configurable for block Copy.

The constraint is that the Rx and Tx channels/rings should be
allocated with same range for it to be used as block copy.

Solver should first assign all the resources to the blockCopy for
all hosts and then rest of the resources should be allocated.

Create anadditional  entry for the same in SYSFW.

Also, format the file for better readability

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>