modules: sysfwResPart: Optimize validation
* Remove check for duplicate host selection and DMSC as host
Now that each host is a dedicated module, its not required.

* Remove the changed state based validation, there are cases
where all instances need to be checked even if one changes.
This fixes the overlap validation error not being shown.

* Remove duplicate code for moduleInstance creation, create
it once as part of validate call and use in others.

* Rewrite the checks for cyclic detection in simple way without
iterating over many modules. Rename array from hostName to hostNames.

* Documentation fixes

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>