scripts: data: Qos.json: Fix groupCount values
Some SoCs represent the QoS base address as the first
register in the endpoint and for others, the base
address has 0x100 added in it.

Adjust the logic to check for both when trying to match
the base address of endpoint from calucated base for a
register.

Also, handle the case of matching with NAVSS IPs where
the design name is "qos_regs" has some prefix.

Regenerate data for all SoCs with this to correct the
groupCount fields in QoS.json

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>