scripts: data: Qos.json: Skip the NAVSS QoS enpoints
NAVSS QoS endpoints are not needed for partitioning
of the peripherals or any other use case.

Modify the parsing script to skip QoS endpoints related
to Main or MCU NAVSS devices.

This will avoid confusion for the user by reducing the
QoS endpoints to only those that he cares about.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>