scripts: parse_resources: Skip the resources which are not available
After finding all the ranges, count the available number of resources.
If the count is 0, skip the resource in the Resources.json

This will make sure that the resource won't show up in the GUI and no
entries will be created for them.

Regenerate data for all platforms after the change.
Only AM64x affects this. Drops un available SAUL channels.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>